Generate FileCheck checks for all module items (#3957)
Instead of only generating checks for functions, generate checks for all named top-level module items, such as types, tags, tables, and memories. Because module items can be in different orders in the input and the output but FileCheck checks must follow the order of the output, we need to be slightly clever about when we emit the checks. Consider these types in the input file:

```
  (type $A (...))
  (type $B (...))
```

If their order is reversed in the output file, then the checks for $B need to be emitted before the checks for $A, so the resulting module will look like this:

```
  ;; CHECK: (type $B (...))
  ;; CHECK: (type $A (...))
  (type $A (...))
  (type $B (...))
```

Rather than this, which looks nicer but would be incorrect:

```
  ;; CHECK: (type $A (...))
  (type $A (...))
  ;; CHECK: (type $B (...))
  (type $B (...))
```