[Wasm GC] Refine function parameter types (#4014)
If a function is always called with a more specific type than it is declared, we can
make the type more specific.

DeadArgumentElimination's name is becoming increasingly misleading, and should
maybe be renamed. But it is the right place for this as it already does an LTO
scan of the call graph and builds up parameter data structures etc.