[Wasm GC] Fix Heap2Local + non-nullable locals (#4017)
Given

(local $x (ref $foo))
(local.set $x ..)
(.. (local.get $x))

If we remove the local.set but not the get, then we end up with

(local $x (ref $foo))
(.. (local.get $x))

It looks like the local.get reads the initial value of a non-nullable local,
which is not allowed.

In practice, this would crash in precompute-propagate which would
try to propagate the initial value to the get. Add an assertion there with
a clear message, as until we have full validation of non-nullable locals
(and the spec for that is in flux), that pass is where bugs will end up
being noticed.

To fix this, replace the get as well. We can replace it with a null
for simplicity; it will never be used anyhow.

This also uncovered a small bug with reached not containing all
the things we reached - it was missing local.gets.