Fix BrOn logic in RemoveUnusedBrs (#4062)
This only moves code around. visitBrOn was in the main part of the pass,
which was incorrect as it could interfere with other work being done there.
Specifically, we have a stack of Ifs there, and if we replace a BrOn with
an If, an assertion was hit. To fix this, run it like sinkBlocks(), in a separate
interleaved phase.

This fixes a bug reported by askeksa-google here:

https://github.com/WebAssembly/gc/issues/226#issuecomment-868739853