Fix signed_ field initialization in Load. (#4075)
This was being set in the creation of Loads in the binary reader, but
forgotten in the SIMD logic - which ends up creating a Load with
type v128, and signed_ was uninitialized.

Very hard to test this, but I saw it "break" hash value computation
which is how I noticed this.

Also initialize the I31 sign field. Now all of them in wasm.h are
properly initialized.