Fix the Switch operand order in LinearExecutionWalker (#4076)
This caused no noticeable bugs, but it could in theory in new passes - in fact
in a pass I will open later this week it did.

Also fix the order in wasm.h. That part has no effect, but it is nice to be
consistent. After this PR, everything should match the single source of
truth which is wasm-delegations-fields.h (as that is used in printing, binary
reading/writing, etc., so it has to be correct). Also Switch now matches
the ordering in Break.