Optimize away dominated calls to functions that run only once (#4111)
Some functions run only once with this pattern:

function foo() {
  if (foo$ran) return;
  foo$ran = 1;
  ...
}

If that global is not ever set to 0, then the function's payload (after the
initial if and return) will never execute more than once. That means we
can optimize away dominated calls:

foo();
foo(); // we can remove this

To do this, we find which globals are "once", which means they can
fit in that pattern, as they are never set to 0. If a function looks like the
above pattern, and it's global is "once", then the function is "once" as
well, and we can perform this optimization.

This removes over 8% of static calls in j2cl.