Select initial contents automatically in fuzzer
This is another attempt to address #4073. Instead of relying on the
timestamp, this examines git log to gather the list of test files added
or modified within some fixed number of days. The number of days is
currently set to 30 (= 1 month) but can be changed. This will be enabled
by `--auto-initial-contents`, which is now disabled by default.

I'm not entirely sure if this will be useful, but uploading this because
this is a simple thing to do. Feedbacks are welcome, and if it turns out
not to be useful, that's also fine.

Fixes (?) #4073.