[Fuzzer] Don't prompt user when seed is given
Fuzzer shows the initial contents and prompts the user if they want to
proceed after #4173, but when the fuzzer is used within a script called
from `wasm-reduce`, we shouldn't pause for the user input. This shows
the prompt only when there is no seed given.

To do that, we now initialize the important initial contents from
`main`. We used to assign those variables before we start `main`.