Detect erroneous use of "FileCheck" in update script (#4198)
The upstream name of the command in LLVM is FileCheck, but we use a Python
package called filecheck instead and using `FileCheck` in lit tests will not
work. To make catching this common error easier, add a warning about it in
scripts/update_lit_checks.py. This warning also applies to other uses of pipes,
which are not supported in this script.