Add runOnModuleCode helper. NFC (#4234)
This method is in parallel to runOnFunction above it. It sets the runner
and then does the walk, like that method.

Also set runner to nullptr by default. I noticed ubsan was warning on
things here, which this should avoid, but otherwise I'm not aware of an
actual bug, so this should be NFC. But it does provide a safer API
that should avoid future bugs.