SimplifyGlobals: Detect trivial read-only-to-write functions (#4257)
We already detected code that looks like

if (foo == 0) {
  foo = 1;
}
That "read only to write" pattern occurs also in functions, like this:

function bar() {
  if (foo == 0) return;
  foo = 1;
}

This PR detects that pattern. It moves code around to share almost
all the logic with the previous pattern (the git diff is not that useful
there, sadly, but looking at them side by side that should be
obvious).

This helps in j2cl on some common clinits, where the clinit function
ends up empty, which is exactly this pattern.